+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - List and delete `DiscoverySettings` objects (API and `replictl`).
 
 ### Changed
+- **BREAKING**: The `events.stream` config block is no longer nested and is now simply `events`.
 - Populate view DB from the events stream.
 - Refactor cluster discovery.
 - Refactor cluster orchestration (also know as refresh).